requirements.txt
PyGithub>=1.54
tqdm>=4.62